atualização

+++ src/entities/Cliente.java
@@ -6,7 +6,7 @@ public class Cliente extends Usuario {
     private String cpf;
 
     public Cliente() {}
-    public Cliente(Long id, String nome, String email, String senha, String cpf) {
+    public Cliente(int id, String nome, String email, String senha, String cpf) {
         super(id, nome, email, senha);
         this.cpf = cpf;
     }
