+++ src/entities/Esterco.java
@@ -2,7 +2,7 @@ package com.src.entities;
 
 public class Esterco extends Produto {
     public Esterco() {}
-    public Esterco(Long id, String nome, double preco, String descricao, int quantidade) {
+    public Esterco(int id, String nome, double preco, String descricao, int quantidade) {
         super(id, nome, preco, descricao, quantidade);
     }
 }
