+++ src/entities/Humus.java
@@ -2,7 +2,7 @@ package com.src.entities;
 
 public class Humus extends Produto {
     public Humus() {}
-    public Humus(Long id, String nome, double preco, String descricao, int quantidade) {
+    public Humus(int id, String nome, double preco, String descricao, int quantidade) {
         super(id, nome, preco, descricao, quantidade);
     }
 }
