+++ src/entities/Muda.java
@@ -2,7 +2,7 @@ package com.src.entities;
 
 public class Muda extends Produto {
     public Muda() {}
-    public Muda(Long id, String nome, double preco, String descricao, int quantidade) {
+    public Muda(int id, String nome, double preco, String descricao, int quantidade) {
         super(id, nome, preco, descricao, quantidade);
     }
 }
